refactor hyperiond
