+++ src/hyperiond/CMakeLists.txt
@@ -1,6 +1,6 @@
 
 add_executable(hyperiond
-		hyperiond.cpp)
+		hyperiond.cpp hyperiond.h main.cpp)
 
 target_link_libraries(hyperiond
 		getoptPlusPlus
